+++ dodSpaceInvaders/src/dod/MemTypes.h
@@ -1,4 +1,5 @@
 #include <cstddef>
+#include <cinttypes>
 
 namespace Dod::MemTypes
 {
@@ -6,6 +7,6 @@ namespace Dod::MemTypes
 	using data_t = std::byte;
 	using dataPoint_t = data_t*;
 	using dataConstPoint_t = const data_t*;
-	using capacity_t = size_t;
+	using capacity_t = int32_t;
 
 };
